+++ saleor/discount/utils.py
@@ -76,8 +76,7 @@ def get_shipping_voucher_discount(voucher, total_price, shipping_price):
 def get_products_voucher_discount(voucher, prices):
     """Calculate discount value for a voucher of product or category type."""
     if voucher.apply_once_per_order:
-        product_total = sum(prices, ZERO_MONEY)
-        return voucher.get_discount_amount_for(product_total)
+        return voucher.get_discount_amount_for(min(prices))
     discounts = (voucher.get_discount_amount_for(price) for price in prices)
     total_amount = sum(discounts, ZERO_MONEY)
     return total_amount