+++ tests/test_checkout.py
@@ -4,6 +4,7 @@ from unittest.mock import Mock, patch
 import pytest
 from django.contrib.auth.models import AnonymousUser
 from django.urls import reverse
+from django.utils import timezone
 from django_countries.fields import Country
 from freezegun import freeze_time
 from prices import Money, TaxedMoney, TaxedMoneyRange
@@ -497,7 +498,7 @@ def test_view_checkout_place_order_with_expired_voucher_code(
     checkout.shipping_method = shipping_zone.shipping_methods.first()
 
     # set voucher to be expired
-    yesterday = datetime.date.today() - datetime.timedelta(days=1)
+    yesterday = timezone.now() - datetime.timedelta(days=1)
     voucher.end_date = yesterday
     voucher.save()
 
@@ -846,6 +847,37 @@ def test_get_discount_for_checkout_value_voucher(
     assert discount == Money(discount_amount, "USD")
 
 
+@pytest.mark.parametrize(
+    "discount_value, discount_type, apply_once_per_order, discount_amount",
+    [
+        (5, DiscountValueType.FIXED, True, 5),
+        (5, DiscountValueType.FIXED, False, 15),
+        (10000, DiscountValueType.FIXED, True, 10),
+        (10, DiscountValueType.PERCENTAGE, True, 1),
+        (10, DiscountValueType.PERCENTAGE, False, 5),
+    ],
+)
+def test_get_discount_for_checkout_apply_once_per_order(
+    checkout_with_items,
+    product_list,
+    discount_value,
+    discount_type,
+    apply_once_per_order,
+    discount_amount,
+):
+    voucher = Voucher.objects.create(
+        code="unique",
+        type=VoucherType.PRODUCT,
+        discount_value_type=discount_type,
+        discount_value=discount_value,
+        apply_once_per_order=apply_once_per_order,
+    )
+    for product in product_list:
+        voucher.products.add(product)
+    discount = get_voucher_discount_for_checkout(voucher, checkout_with_items)
+    assert discount == Money(discount_amount, "USD")
+
+
 def test_get_discount_for_checkout_value_voucher_not_applicable():
     voucher = Voucher(
         code="unique",
@@ -884,6 +916,7 @@ def test_get_discount_for_checkout_shipping_voucher(
         get_subtotal=Mock(return_value=subtotal),
         is_shipping_required=Mock(return_value=True),
         shipping_method=Mock(get_total=Mock(return_value=shipping_total)),
+        get_shipping_price=Mock(return_value=shipping_total),
         shipping_address=Mock(country=Country(shipping_country_code)),
     )
     voucher = Voucher(
@@ -904,6 +937,7 @@ def test_get_discount_for_checkout_shipping_voucher_all_countries():
         get_subtotal=Mock(return_value=subtotal),
         is_shipping_required=Mock(return_value=True),
         shipping_method=Mock(get_total=Mock(return_value=shipping_total)),
+        get_shipping_price=Mock(return_value=shipping_total),
         shipping_address=Mock(country=Country("PL")),
     )
     voucher = Voucher(
@@ -1000,6 +1034,7 @@ def test_get_discount_for_checkout_shipping_voucher_not_applicable(
         get_subtotal=Mock(return_value=subtotal),
         is_shipping_required=Mock(return_value=is_shipping_required),
         shipping_method=shipping_method,
+        get_shipping_price=Mock(return_value=Money(10, "USD")),
     )
     voucher = Voucher(
         code="unique",
@@ -1080,7 +1115,7 @@ def test_checkout_voucher_form_active_queryset_voucher_not_active(
     voucher, request_checkout_with_item
 ):
     assert Voucher.objects.count() == 1
-    voucher.start_date = datetime.date.today() + datetime.timedelta(days=1)
+    voucher.start_date = timezone.now() + datetime.timedelta(days=1)
     voucher.save()
     form = CheckoutVoucherForm(
         {"voucher": voucher.code}, instance=request_checkout_with_item
@@ -1093,7 +1128,7 @@ def test_checkout_voucher_form_active_queryset_voucher_active(
     voucher, request_checkout_with_item
 ):
     assert Voucher.objects.count() == 1
-    voucher.start_date = datetime.date.today()
+    voucher.start_date = timezone.now()
     voucher.save()
     form = CheckoutVoucherForm(
         {"voucher": voucher.code}, instance=request_checkout_with_item
@@ -1106,8 +1141,8 @@ def test_checkout_voucher_form_active_queryset_after_some_time(
     voucher, request_checkout_with_item
 ):
     assert Voucher.objects.count() == 1
-    voucher.start_date = datetime.date(year=2016, month=6, day=1)
-    voucher.end_date = datetime.date(year=2016, month=6, day=2)
+    voucher.start_date = timezone.now().replace(year=2016, month=6, day=1, hour=0)
+    voucher.end_date = timezone.now().replace(year=2016, month=6, day=2, hour=0)
     voucher.save()
 
     with freeze_time("2016-05-31"):
@@ -1116,7 +1151,7 @@ def test_checkout_voucher_form_active_queryset_after_some_time(
         )
         assert form.fields["voucher"].queryset.count() == 0
 
-    with freeze_time("2016-06-01"):
+    with freeze_time("2016-06-01T05:00:00+00:00"):
         form = CheckoutVoucherForm(
             {"voucher": voucher.code}, instance=request_checkout_with_item
         )
@@ -1135,7 +1170,7 @@ def test_get_voucher_for_checkout(checkout_with_voucher, voucher):
 
 
 def test_get_voucher_for_checkout_expired_voucher(checkout_with_voucher, voucher):
-    date_yesterday = datetime.date.today() - datetime.timedelta(days=1)
+    date_yesterday = timezone.now() - datetime.timedelta(days=1)
     voucher.end_date = date_yesterday
     voucher.save()
     checkout_voucher = get_voucher_for_checkout(checkout_with_voucher)
@@ -1171,6 +1206,15 @@ def test_recalculate_checkout_discount(
     assert checkout_with_voucher.discount_amount == Money("10.00", "USD")
 
 
+def test_recalculate_checkout_discount_with_sale(
+    checkout_with_voucher_percentage, discount_info
+):
+    checkout = checkout_with_voucher_percentage
+    recalculate_checkout_discount(checkout, [discount_info])
+    assert checkout.discount_amount == Money("1.50", "USD")
+    assert checkout.get_total(discounts=[discount_info]) == Money("13.50", "USD")
+
+
 def test_recalculate_checkout_discount_voucher_not_applicable(
     checkout_with_voucher, voucher
 ):
@@ -1187,7 +1231,7 @@ def test_recalculate_checkout_discount_voucher_not_applicable(
 
 def test_recalculate_checkout_discount_expired_voucher(checkout_with_voucher, voucher):
     checkout = checkout_with_voucher
-    date_yesterday = datetime.date.today() - datetime.timedelta(days=1)
+    date_yesterday = timezone.now() - datetime.timedelta(days=1)
     voucher.end_date = date_yesterday
     voucher.save()