+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable, unreleased changes to this project will be documented in this file.
 - A few unused panels are now disabled by default from the debug toolbar; this should improve loading time when debugging - #4301 by @NyanKiyoshi
 - Fixed internal error when adding a note to an anonymous order - #4319 by @NyanKiyoshi
 - Fix populatedb gift card duplication error - #4336 by @fowczarek
+- Fix voucher apply once per order - #4339 by @fowczarek
+- Change type of start and end date for discounts on date time field - #4293 by @fowczarek
 
 ## 2.7.0
 
@@ -71,7 +73,6 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fix margin calculations when product/variant price is set to zero - #4170 by @MahmoudRizk
 - Fix applying discounts in checkout's subtotal calculation in API - #4192 by @maarcingebala
 - Fix GATEWAYS_ENUM to always contain all implemented payment gateways - #4108 by @koradon
-
 
 ## 2.6.0